CAR-29 создал стор для игры
