+++ packages/client/src/shared/store/game/types.ts
@@ -1,0 +1,3 @@
+import type { characters } from '@/shared/constants/options'
+
+export type TCharacters = keyof typeof characters | null
